feat: extract table state hook and toolbar
